migration update

+++ database/migrations/2022_11_22_102234_create_post_tag_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
     {
         Schema::create('post_tag', function (Blueprint $table) {
             $table->primary(['post_id', 'tag_id']);
-            $table->unsignedInteger('post_id');
-            $table->unsignedInteger('tag_id');
+            $table->unsignedBigInteger('post_id');
+            $table->unsignedBigInteger('tag_id');
             $table->timestamps();
 
             $table->foreign('post_id')->references('id')->on('posts')->onDelete('cascade');
